recommend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
scipy==1.14.0
Werkzeug==3.0.3
python-dotenv==1.0.1